refactor: use~Input으로 네이밍 수정
